+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/configuration/SpecificationArgumentsResolverConfig.java
@@ -19,6 +19,7 @@
 
 package de.frachtwerk.essencium.backend.configuration;
 
+import de.frachtwerk.essencium.backend.configuration.properties.EssenciumJpaProperties;
 import de.frachtwerk.essencium.backend.controller.access.AccessAwareSpecArgResolver;
 import de.frachtwerk.essencium.backend.model.AbstractBaseUser;
 import de.frachtwerk.essencium.backend.model.dto.EssenciumUserDetailsImpl;
@@ -43,10 +44,12 @@ public class SpecificationArgumentsResolverConfig<
     implements WebMvcConfigurer {
   private final AbstractApplicationContext applicationContext;
   private final AbstractUserService<USER, AUTHUSER, ID, T> userService;
+  private final EssenciumJpaProperties essenciumJpaProperties;
 
   @Override
   public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
-    argumentResolvers.add(new AccessAwareSpecArgResolver(applicationContext, userService));
+    argumentResolvers.add(
+        new AccessAwareSpecArgResolver(applicationContext, userService, essenciumJpaProperties));
   }
 
   @Override